requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytz==2024.1
redis==5.0.2
requests==2.31.0
service-identity==24.1.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.7.0
websockets==12.0
zope.interface==6.2
